server/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
python-dotenv
alembic
pydantic
